requirements.txt
Flask==2.1.2
flask-restx==0.5.1
flask_jwt_extended==4.4.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.1.2
pytest==7.1.2
python-dotenv==0.10.1
xlsxwriter
asyncio
boto3
Faker
uuid
backoff
simple_cache
requests
